requirements.txt
langchain
langchain-openai
langchain-community
ollama
crewai
openai
faiss-cpu
streamlit
pypdf